FIX: Fixed sparse to banded conversion
Only relevant for SciPy versions 1.15 and newer, since previously pybaselines would use the private SciPy method scipy.interpolate._bspl._norm_eq_lsq for setting up penalized spline solvers. Corrected the retrieval of banded data from sparse matrices, which would often truncate columns with zeros and cause issues. Also removed _spline_utils._solve_pspline since it was replaced by _spline_utils.PSpline.solve_pspline.

+++ pybaselines/two_d/_whittaker_utils.py
@@ -84,7 +84,7 @@ class PenalizedSystem2D:
         ----------
         data_size : Sequence[int, int]
             The number of data points for the system.
-        lam : float or Sequence[int, int], optional
+        lam : float or Sequence[float, float], optional
             The penalty factor applied to the difference matrix for the rows and columns,
             respectively. If a single value is given, both will use the same value. Larger
             values produce smoother results. Must be greater than 0. Default is 1.
